+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Added
 
 - **! Added `confz` as a dependency!**
+- Added `fail_fast` to config
 
 ## [0.3.0] - 2023-02-25